Update BlockMover stories
